test(S8-step2): 建 transfer 共享工厂 + make_queue_chain（安全网奠基）
S8-step2 第一刀（test-only，零生产改动）：抽取 TransferService 前的特征化测试安全网奠基。

- 新增 tests/transfer_fixtures.py：make_transfer_chain/make_task/make_media_info/
  FakeMeta/FakeMedia/make_fileitem/migrate_to_media_job 单一来源（此前在
  test_transfer_job_manager.py 与 test_transfer_handle_integration.py 重复），
  两文件改为 `from tests.transfer_fixtures import` 去重。
- 新增 make_queue_chain()：在 make_transfer_chain 基础上种入队列/守护/计数器机器属性
  （_queue/_threads/_queue_active/计数器/_progress），不 spawn 线程 → 让 worker loop
  (__start_transfer) 可在 venv 内被受控驱动（真实 TransferChain() 会起线程阻塞 15s）。
- 新增 tests/test_transfer_queue_machinery.py：make_queue_chain 夹具契约测试
  （Step 2 worker-loop 特征化测试的归宿）。
- 全套 transfer 测试 67 passed。
